+++ packages/php-wasm/universal/src/lib/index.ts
@@ -79,7 +79,7 @@ export {
 } from './urls';
 
 export { isExitCode } from './is-exit-code';
-export { proxyFileSystem } from './proxy-file-system';
+export { proxyFileSystem, isPathToSharedFS } from './proxy-file-system';
 export { sandboxedSpawnHandlerFactory } from './sandboxed-spawn-handler-factory';
 
 export * from './api';