+++ packages/playground/blueprints/src/lib/steps/activate-plugin.spec.ts
@@ -8,13 +8,13 @@ import { activatePlugin } from './activate-plugin';
 import { phpVar } from '@php-wasm/util';
 import type { PHPRequestHandler } from '@php-wasm/universal';
 import { loadNodeRuntime } from '@php-wasm/node';
-import { bootWordPress } from '@wp-playground/wordpress';
+import { bootWordPressAndRequestHandler } from '@wp-playground/wordpress';
 
 describe('Blueprint step activatePlugin()', () => {
 	let php: PHP;
 	let handler: PHPRequestHandler;
 	beforeEach(async () => {
-		handler = await bootWordPress({
+		handler = await bootWordPressAndRequestHandler({
 			createPhpRuntime: async () =>
 				await loadNodeRuntime(RecommendedPHPVersion),
 			siteUrl: 'http://playground-domain/',