+++ packages/playground/blueprints/src/lib/v2/run-blueprint-v2.ts
@@ -6,10 +6,11 @@ import {
 import { phpVar } from '@php-wasm/util';
 import { getV2Runner } from './get-v2-runner';
 import {
-	type BlueprintV2Declaration,
-	type ParsedBlueprintV2Declaration,
+	type RawBlueprintV2Data,
+	type ParsedBlueprintV1orV2String,
 	parseBlueprintDeclaration,
 } from './blueprint-v2-declaration';
+import type { BlueprintV1Declaration } from '../v1/types';
 
 export type PHPExceptionDetails = {
 	exception: string;
@@ -32,7 +33,10 @@ export type BlueprintMessage =
 interface RunV2Options {
 	php: UniversalPHP;
 	cliArgs?: string[];
-	blueprint: BlueprintV2Declaration | ParsedBlueprintV2Declaration;
+	blueprint:
+		| RawBlueprintV2Data
+		| ParsedBlueprintV1orV2String
+		| BlueprintV1Declaration;
 	blueprintOverrides?: {
 		wordpressVersion?: string;
 		additionalSteps?: any[];