+++ packages/playground/website/src/components/ensure-playground-site/ensure-playground-site-is-selected.tsx
@@ -1,6 +1,4 @@
 import { useEffect, useState } from 'react';
-import type { ResolvedBlueprint } from '../../lib/state/url/resolve-blueprint-from-url';
-import { resolveBlueprintFromURL } from '../../lib/state/url/resolve-blueprint-from-url';
 import { useCurrentUrl } from '../../lib/state/url/router-hooks';
 import { opfsSiteStorage } from '../../lib/state/opfs/opfs-site-storage';
 import {
@@ -21,6 +19,7 @@ import { usePrevious } from '../../lib/hooks/use-previous';
 import { modalSlugs } from '../layout';
 import { setActiveModal } from '../../lib/state/redux/slice-ui';
 import { selectClientBySiteSlug } from '../../lib/state/redux/slice-clients';
+import { randomSiteName } from '../../lib/state/redux/random-site-name';
 
 /**
  * Ensures the redux store always has an activeSite value.
@@ -151,50 +150,16 @@ export function EnsurePlaygroundSiteIsSelected({
 	return children;
 }
 
-function parseSearchParams(searchParams: URLSearchParams) {
-	const params: Record<string, any> = {};
-	for (const key of searchParams.keys()) {
-		const value = searchParams.getAll(key);
-		params[key] = value.length > 1 ? value : value[0];
-	}
-	return params;
-}
 async function createNewTemporarySite(
 	dispatch: ReturnType<typeof useAppDispatch>,
 	requestedSiteSlug?: string
 ) {
 	// If the site slug is missing, create a new temporary site.
-	// Lean on the Query API parameters and the Blueprint API to
-	// create the new site.
-	const newUrl = new URL(window.location.href);
-	const defaultBlueprint =
-		'https://raw.githubusercontent.com/WordPress/blueprints/refs/heads/trunk/blueprints/welcome/blueprint.json';
-	let resolvedBlueprint: ResolvedBlueprint | undefined = undefined;
-
-	try {
-		resolvedBlueprint = await resolveBlueprintFromURL(
-			newUrl,
-			defaultBlueprint
-		);
-	} catch (e) {
-		logger.error('Error resolving blueprint:', e);
-	}
-
-	// Create a new site otherwise
+	const siteName = requestedSiteSlug
+		? deriveSiteNameFromSlug(requestedSiteSlug)
+		: randomSiteName();
 	const newSiteInfo = await dispatch(
-		setTemporarySiteSpec({
-			metadata: {
-				originalBlueprint: resolvedBlueprint?.blueprint,
-				originalBlueprintSource: resolvedBlueprint?.source,
-				name: requestedSiteSlug
-					? deriveSiteNameFromSlug(requestedSiteSlug)
-					: undefined,
-			},
-			originalUrlParams: {
-				searchParams: parseSearchParams(newUrl.searchParams),
-				hash: newUrl.hash,
-			},
-		})
+		setTemporarySiteSpec(siteName, new URL(window.location.href))
 	);
 	await dispatch(setActiveSite(newSiteInfo.slug));
 }